refactor(apps/dashboard): replace transition with blur-fade
